Add files via upload
I created a project for my programming II class (COP 3337) at Florida International University. This project was based on creating a simple banking application where the user can create a checking and savings account. The two actions to be performed in this project are withdrawing or depositing money to a new account. This project was assigned to us as practice for learning the Inheritance concepts.